Add files via upload
Regarding the 6 most recent uploads; We are having some trouble generating random questions and answers. Put simply, when a question is answered, the answer stays the same so all other questions are wrong. We think this is due to a problem with calling back answers from another matlab script. Any help or a point in the right direction would be much appreciated